net: phy: swconfig: add switch layer
Add the switch configuration API based on netlink.
This allows to configure the VLAN features of certain switches.

We need this driver because of the ag71xx driver dependency.

This patch is based on the OpenWrt implementation.
Because this implementation is an OpenWrt hack we have to use Switchdev
instead.

Signed-off-by: Rocco Folino <rocco.folino@tanaza.com>