net: phy: add detach callback to struct phy_driver
Add detach callback to the struct phy_driver. The callback is called
before an ethernet device is detached from the PHY.

This patch is based on the OpenWrt implementation.
This patch is needed because the ar8216/ar8327 driver depends on it.
We should remove this patch in the future and to port the depending drivers
to be upstream compliant.

Signed-off-by: Gabor Juhos <juhosg@openwrt.org>
Signed-off-by: Rocco Folino <rocco.folino@tanaza.com>